requirements.txt
# Confluence Content Copier - Dependencies

# HTTP requests
requests>=2.28.0

# YAML parsing for configuration
PyYAML>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
responses>=0.22.0  # For mocking requests
